// src/api/generated/models/chatConversation.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Shadcn Admin API
 * API for the Shadcn Admin dashboard application
 * OpenAPI spec version: 1.0.0
 */
import type { ChatMessage } from './chatMessage';

export interface ChatConversation {
  id: string;
  username: string;
  fullName: string;
  title?: string;
  profile?: string;
  messages: ChatMessage[];
}