// src/api/generated/models/taskStatus.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Shadcn Admin API
 * API for the Shadcn Admin dashboard application
 * OpenAPI spec version: 1.0.0
 */

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TaskStatus = {
  todo: 'todo',
  in_progress: 'in progress',
  done: 'done',
  canceled: 'canceled',
  backlog: 'backlog',
} as const;