// src/api/generated/models/taskLabel.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Shadcn Admin API
 * API for the Shadcn Admin dashboard application
 * OpenAPI spec version: 1.0.0
 */

export type TaskLabel = typeof TaskLabel[keyof typeof TaskLabel];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TaskLabel = {
  bug: 'bug',
  feature: 'feature',
  documentation: 'documentation',
} as const;