// src/api/generated/models/taskPriority.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Shadcn Admin API
 * API for the Shadcn Admin dashboard application
 * OpenAPI spec version: 1.0.0
 */

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;